+++ src/Build-Latex.ps1
@@ -4,9 +4,10 @@ function Build-Latex
         [string]$OutputDirectory = "..\out",
         [Boolean]$ShouldCompileBibtex = $true
     )
-        $InitialDirectory = Get-Location
 
-        $ErrorActionPreference = "Stop"
+    try
+    {
+        $InitialDirectory = Get-Location
 
         Write-Host "Initial working directory: $InitialDirectory"  -ForegroundColor Magenta
         Write-Host "Setting working directory to PSScriptRoot: $PSScriptRoot"  -ForegroundColor Magenta
@@ -66,13 +67,19 @@ function Build-Latex
 
         Set-Location $InitialDirectory
 
-        $EncodingScriptPath = "$InitialDirectory\scripts\Test-Encoding.ps1"
+        $EncodingScriptPath = "$InitialDirectory/scripts/Test-Encoding.ps1"
 
         & $EncodingScriptPath -Autofix
 
         Write-Host "Fix encoding is complete." -ForegroundColor Green
         Write-Host "Exit Code: $LASTEXITCODE" -ForegroundColor Green
         Write-Host "Changing Powershell Directory to $InitialDirectory ... " -ForegroundColor Green
+    }
+    catch
+    {
+        Set-Location $InitialDirectory
+        Write-Error "Latex error appeared. Inspect the logs."
+    }
 }
 
 function Write-Dash